Copilot makes build.sh less verbose
---
Lets reduce the output of ./build.sh in the case everything is fine.

Please run it once to get the feel.

It could work by directing the output to some .gitignored build/* files and only saying to check those files if some step fails.

Not sure if build/ dir is idiomatic name to use in a python project so please use something else if it makes sense.