Copilot refactored test_integration.py to remove ide error
---

IDE is showing these errors:
```
Type of "backup_pairs" is partially unknown
  Type of "backup_pairs" is "list[dict[str, Unknown]]"
```

But ./build.sh is happy. Do you know what is going on?
---
Ah... note it is the same dataclass that is in config.py.

I think we should use that and modify test_utils.py and test_integration.py just use that.

---
Please run ./build.sh instead to get linters etc