Copilot created a red test for a suspected issue
---

Please study the files provided.

I think there may be an issue with the backup logic.

However, lets just define _failing tests first_ to kind of spec out what we want.

The problem I anticipate is that there are snapshots in the source that are older than the newest snapshot in the target, but they don't exist in the target.

In this case we should log a warning, but not attempt to send them to the target.

The copied snapshots should always follow the same sequence of "parenting" -p
Attempting to send some straggler snapshot from source creates the issue that the "chain of parenting" will be non sequential.

Can you create a new integration test that tries to capture this. Ie source has extra snapshots that shouldn't be copied and don't fit to the existing "chain".

I'm not sure I'm being clear with this. Please ask for clarification if needed.

Please follow an identical pattern used with testing other backup functionality.

Please use `./build.sh` and `./update-test-references.py -y` as needed.

To save context space and keep the chat compact:

- Only do the file changes
- Don't describe before or after what is being done